@@ -3,16 +3,16 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"hamburger-react": "2.4.1",
+		"hamburger-react": "2.5.1",
 		"i18next": "21.6.13",
 		"i18next-browser-languagedetector": "6.1.3",
-		"react": "17.0.2",
-		"react-dom": "17.0.2",
+		"react": "18.3.1",
+		"react-dom": "18.3.1",
 		"react-i18next": "11.15.6",
-		"react-scripts": "5.0.0",
-		"styled-components": "5.3.3",
-		"typescript": "4.4.2",
-		"web-vitals": "2.1.0"
+		"react-scripts": "5.0.1",
+		"styled-components": "6.1.13",
+		"typescript": "5.7.2",
+		"web-vitals": "4.2.4"
 	},
 	"scripts": {
 		"start": "react-scripts start",
@@ -41,27 +41,27 @@
 		]
 	},
 	"devDependencies": {
-		"@commitlint/cli": "16.2.1",
-		"@commitlint/config-conventional": "16.2.1",
-		"@testing-library/jest-dom": "5.16.2",
-		"@testing-library/react": "12.1.2",
-		"@testing-library/user-event": "13.5.0",
-		"@types/jest": "27.4.0",
-		"@types/node": "17.0.18",
-		"@types/react": "17.0.39",
-		"@types/react-dom": "17.0.11",
-		"@types/styled-components": "5.1.22",
-		"@typescript-eslint/eslint-plugin": "5.12.0",
-		"@typescript-eslint/parser": "5.12.0",
-		"eslint": "8.9.0",
+		"@commitlint/cli": "19.6.0",
+		"@commitlint/config-conventional": "19.6.0",
+		"@testing-library/jest-dom": "6.6.3",
+		"@testing-library/react": "16.0.1",
+		"@testing-library/user-event": "14.5.2",
+		"@types/jest": "29.5.14",
+		"@types/node": "22.9.4",
+		"@types/react": "18.3.12",
+		"@types/react-dom": "18.3.1",
+		"@types/styled-components": "5.1.34",
+		"@typescript-eslint/eslint-plugin": "8.16.0",
+		"@typescript-eslint/parser": "8.16.0",
+		"eslint": "9.15.0",
 		"eslint-config-airbnb": "19.0.4",
-		"eslint-config-airbnb-typescript": "16.1.0",
-		"eslint-config-prettier": "8.3.0",
-		"eslint-plugin-import": "2.25.4",
-		"eslint-plugin-jsx-a11y": "6.5.1",
-		"eslint-plugin-prettier": "4.0.0",
-		"eslint-plugin-react": "7.28.0",
-		"husky": "7.0.4",
-		"prettier": "2.5.1"
+		"eslint-config-airbnb-typescript": "18.0.0",
+		"eslint-config-prettier": "9.1.0",
+		"eslint-plugin-import": "2.31.0",
+		"eslint-plugin-jsx-a11y": "6.10.2",
+		"eslint-plugin-prettier": "5.2.1",
+		"eslint-plugin-react": "7.37.2",
+		"husky": "9.1.7",
+		"prettier": "3.3.3"
 	}
 }
