@@ -5,7 +5,29 @@ export const ocupation = {
 		{
 			id: 1,
 			title: 'Desenvolvedor Front-End',
-			fromTo: 'mai. de 2021 – o momento',
+			fromTo: 'dez. de 2023 – o momento',
+			company: 'Arcotech',
+			description: [
+				{
+					id: 1,
+					text: `Utilizamos a metodologia Shape Up para o planejamento das sprints. Na maioria dos
+					projetos, seguimos o fluxo Trunk-Based Development; contudo, há algumas exceções em que
+					adotamos o GitFlow. Além disso, trabalho com diversas ferramentas de monitoramento, como
+					ArgoCD, Analytics, Grafana e New Relic.`,
+				},
+				{
+					id: 2,
+					text: `No dia a dia, utilizo React com TypeScript e Zustand para o gerenciamento de estados da aplicação.
+					Também trabalhamos com micro front-end utilizando module federation. Para testes unitários,
+					usamos Jest, e no backend for frontend (BFF), utilizamos NestJS. Nossas aplicações estão hospedadas
+					na AWS, com o CloudFront como CDN e Kubernetes para o gerenciamento dos containers Docker.`,
+				},
+			],
+		},
+		{
+			id: 2,
+			title: 'Desenvolvedor Front-End',
+			fromTo: 'mai. de 2021 – dez. 2023 - 2 anos e 8 meses',
 			company: 'Zenvia',
 			description: [
 				{
@@ -19,13 +41,12 @@ export const ocupation = {
 					text: `Durante o dia-a-dia eu utilizo: NextJS + TypeScript + Redux Saga + Context API + Styled Components,
 						para testes usamos Cypress(e2e), Jest(unitários), também implementamos jsDocs para comentar o objetivo das funções,
 						aplicamos o conventional commits para manter os commits sucintos, eu também cuido de algumas libs,
-						De componentes, tokens, ícones... Eu também do Hotjar/Google Analytics junto com o resto do time de
-						Front`,
+						De componentes, tokens, ícones... Usamos Hotjar e Google Analytics para monitoria dos projetos`,
 				},
 			],
 		},
 		{
-			id: 2,
+			id: 3,
 			title: 'Desenvolvedor Front-End',
 			fromTo: 'fev. de 2021 – ago. de 2021 - 7 meses',
 			company: 'NovaHaus',
@@ -48,7 +69,7 @@ export const ocupation = {
 			],
 		},
 		{
-			id: 3,
+			id: 4,
 			title: 'Desenvolvedor Full Stack',
 			fromTo: 'set de 2020 - jan de 2021 - 5 meses',
 			company: 'NWB - Network Brasil',
